feat: export event photos, hook refactor
